Add the chain spec
