+++ node/src/parachain/dev_chain_spec.rs
@@ -1,0 +1,158 @@
+use peaq_dev_runtime::{
+	AccountId, BalancesConfig, EVMConfig, EthereumConfig, GenesisAccount, GenesisConfig,
+	Signature, SudoConfig, SystemConfig, WASM_BINARY, Precompiles, ParachainInfoConfig,
+};
+use sc_service::{ChainType, Properties};
+use sp_consensus_aura::sr25519::AuthorityId as AuraId;
+use sp_core::{sr25519, Pair, Public};
+use sp_runtime::traits::{IdentifyAccount, Verify};
+use cumulus_primitives_core::ParaId;
+use crate::parachain::Extensions;
+
+/// Specialized `ChainSpec`. This is a specialization of the general Substrate ChainSpec type.
+pub type ChainSpec = sc_service::GenericChainSpec<GenesisConfig, Extensions>;
+
+/// Generate a crypto pair from seed.
+pub fn get_from_seed<TPublic: Public>(seed: &str) -> <TPublic::Pair as Pair>::Public {
+	TPublic::Pair::from_string(&format!("//{}", seed), None)
+		.expect("static values are valid; qed")
+		.public()
+}
+
+type AccountPublic = <Signature as Verify>::Signer;
+
+/// Generate an account ID from seed.
+pub fn get_account_id_from_seed<TPublic: Public>(seed: &str) -> AccountId
+where
+	AccountPublic: From<<TPublic::Pair as Pair>::Public>,
+{
+	AccountPublic::from(get_from_seed::<TPublic>(seed)).into_account()
+}
+
+/// Generate an Aura authority key.
+pub fn authority_keys_from_seed(s: &str) -> (AccountId, AuraId) {
+	(get_account_id_from_seed::<sr25519::Public>(s), get_from_seed::<AuraId>(s))
+}
+
+pub fn development_config(para_id: u32) -> Result<ChainSpec, String> {
+	let wasm_binary = WASM_BINARY.ok_or_else(|| "Development wasm not available".to_string())?;
+
+	let mut properties = Properties::new();
+	properties.insert("tokenSymbol".into(), "PEAQ".into());
+	properties.insert("tokenDecimals".into(), 18.into());
+
+	Ok(ChainSpec::from_genesis(
+		"peaq-dev",
+		"peaq-substrate-dev",
+		ChainType::Development,
+		move || {
+			configure_genesis(
+				wasm_binary,
+				// Initial PoA authorities
+				vec![
+					authority_keys_from_seed("Alice"),
+				],
+				// Sudo account
+				get_account_id_from_seed::<sr25519::Public>("Alice"),
+				// Pre-funded accounts
+				vec![
+					get_account_id_from_seed::<sr25519::Public>("Alice"),
+					get_account_id_from_seed::<sr25519::Public>("Bob"),
+					get_account_id_from_seed::<sr25519::Public>("Alice//stash"),
+					get_account_id_from_seed::<sr25519::Public>("Bob//stash"),
+					get_account_id_from_seed::<sr25519::Public>("Charlie"),
+				],
+				para_id.into(),
+			)
+		},
+		// Bootnodes
+		vec![],
+		// Telemetry
+		None,
+		// Protocol ID
+		None,
+		// Fork ID
+		None,
+		// Properties
+		Some(properties),
+		// Extensions
+		Extensions {
+			bad_blocks: Default::default(),
+			relay_chain: "test-service".into(),
+			para_id: para_id,
+		},
+	))
+}
+
+pub fn session_keys(aura: AuraId) -> peaq_dev_runtime::opaque::SessionKeys {
+    peaq_dev_runtime::opaque::SessionKeys { aura }
+}
+
+/// Configure initial storage state for FRAME modules.
+fn configure_genesis(
+	wasm_binary: &[u8],
+	initial_authorities: Vec<(AccountId, AuraId)>,
+	root_key: AccountId,
+	endowed_accounts: Vec<AccountId>,
+    parachain_id: ParaId,
+) -> GenesisConfig {
+	// This is supposed the be the simplest bytecode to revert without returning any data.
+	// We will pre-deploy it under all of our precompiles to ensure they can be called from
+	// within contracts.
+	// (PUSH1 0x00 PUSH1 0x00 REVERT)
+	let revert_bytecode = vec![0x60, 0x00, 0x60, 0x00, 0xFD];
+
+	GenesisConfig {
+		system: SystemConfig {
+			// Add Wasm runtime to storage.
+			code: wasm_binary.to_vec(),
+		},
+		parachain_info: ParachainInfoConfig { parachain_id },
+		balances: BalancesConfig {
+			// Configure endowed accounts with initial balance of 1 << 78.
+			balances: endowed_accounts
+				.iter()
+				.cloned()
+				.map(|k| (k, 1 << 78))
+				.collect(),
+		},
+		// [TODO]...
+		// block_reward
+		session: peaq_dev_runtime::SessionConfig {
+			keys: initial_authorities
+				.iter()
+				.map(|x| (x.0.clone(), x.0.clone(), session_keys(x.1.clone())))
+				.collect::<Vec<_>>(),
+		},
+		aura: Default::default(),
+		sudo: SudoConfig {
+			// Assign network admin rights.
+			key: Some(root_key),
+		},
+		aura_ext: Default::default(),
+		collator_selection: peaq_dev_runtime::CollatorSelectionConfig {
+			desired_candidates: 200,
+			// [TODO]...
+			candidacy_bond: 32_000,
+			invulnerables: initial_authorities.iter().map(|x| x.0.clone()).collect::<Vec<_>>(),
+		},
+		evm: EVMConfig {
+			accounts: Precompiles::used_addresses()
+				.map(|addr| {
+					(
+						addr.into(),
+						GenesisAccount {
+							nonce: Default::default(),
+							balance: Default::default(),
+							storage: Default::default(),
+							code: revert_bytecode.clone(),
+						},
+					)
+				})
+				.collect(),
+		},
+		ethereum: EthereumConfig {},
+		dynamic_fee: Default::default(),
+		base_fee: Default::default(),
+	}
+}
